requirements.txt
pymongo~=4.4.1
python-dotenv~=1.0.0
pydantic~=2.1.1
fastapi~=0.101.1